requirements.txt
albumentations==1.1.0
einops==0.3.2
munch==2.5.0
numpy==1.21.4
#opencv_python==4.5.4.60
pandas==1.3.5
pdf2image==1.16.0
Pillow==9.0.0
PyYAML==6.0
streamlit==1.3.0
timm==0.4.12
torch==1.10.1
torchvision==0.11.2
transformers==4.45.0
protobuf==3.20.0
#wandb==0.12.9
x_transformers
opencv-python-headless
gdown
altair<5.0.0